python/requirements.txt
pandas
psycopg2
python-dotenv
requests
websocket-client
logging
asyncio
